+++ addon/components/flexberry-dropdown.js
@@ -47,7 +47,7 @@ export default FlexberryBaseComponent.extend({
     @property placeholder
     @type String
     @default t('components.flexberry-dropdown.placeholder')
-   */
+  */
   placeholder: t('components.flexberry-dropdown.placeholder'),
 
   /**
@@ -63,10 +63,10 @@ export default FlexberryBaseComponent.extend({
     Available items.
 
     @property items
-    @type Array
-    @default []
+    @type Array|Object
+    @default {}
   */
-  items: [],
+  items: {},
 
   /**
     DOM-element representing semantic ui-dropdown component.
@@ -114,26 +114,39 @@ export default FlexberryBaseComponent.extend({
   /**
     Handles changes in available items & selected item (including changes on component initialization).
   */
-  itemsOrValueDidChange: Ember.on('init', Ember.observer('items.[]', 'value', function() {
+  itemsOrValueDidChange: Ember.on('init', Ember.observer('items', 'value', function() {
     let destroyHasBeenCalled = this.get('destroyHasBeenCalled');
     if (destroyHasBeenCalled) {
       return;
     }
 
     let items = this.get('items');
-    let needChecksOnValue = this.get('needChecksOnValue');
-    if (needChecksOnValue && !Ember.isArray(items)) {
-      Ember.Logger.error(`Wrong type of flexberry-dropdown \`items\` propery: actual type is ${Ember.typeOf(items)}, but array is expected.`);
+    
+    if (Ember.isArray(items)) {
+      let obj = {};
+
+      items.forEach(function(item) {
+        obj[item] = item.toString();
+      });
+
+      this.set('items', obj);
     }
+
+    let needChecksOnValue = this.get('needChecksOnValue');
 
     // Convert 'value' and 'items' to strings because flexberry-dropdown interpret selected value as string commonly.
     let value = this.get('value') || null;
     let stringValue = value ? value.toString() : null;
-    let itemsString = items.map(function(item) {
-      return item ? item.toString() : null;
-    });
 
-    if (needChecksOnValue && !Ember.isNone(stringValue) && itemsString.indexOf(stringValue) < 0) {
+    let valueIsExist = false;
+    for (let key in items) {
+      if (items[key] === stringValue) {
+        valueIsExist = true;
+        break;
+      }
+    }
+
+    if (needChecksOnValue && !Ember.isNone(stringValue) && !valueIsExist) {
       Ember.Logger.error(`Wrong value of flexberry-dropdown \`value\` propery: \`${value}\`. Allowed values are: [\`${items.join(`\`, \``)}\`].`);
     }