+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Blueprints:
     * Add localization support for generated ember entities.
+* `flexberry-dropdown` component:
+    * Add support object type for `items` property.
 
 ## [0.3.0] - 2016-07-29
 ### Added